XP : store the RNG raw binary data for each xp
It allows us to guaranty that user didn't manipulate the data.
It's also usefull to test data randomness.
The downfall is that it will use a lot more space (~400kb per xp)...

We change the xp protocole and now it's the RNG which close the ws connection, when it has uploaded the raw data.
And the client remove itself from queue at the moment and not before.